Folders created for project, create tests for dice and craps, start working on player and dealer
