--- src/main/java/com/codedifferently/casino/BlackJack.java
@@ -1,3 +1,0 @@
-public class BlackJack {
-  
-}
