--- src/main/java/com/codedifferently/casino/Card.java
@@ -1,3 +1,0 @@
-public class Card {
-  
-}
