--- src/main/java/com/codedifferently/casino/CardGame.java
@@ -1,3 +1,0 @@
-public class CardGame {
-  
-}
