--- src/main/java/com/codedifferently/casino/Deck.java
@@ -1,3 +1,0 @@
-public class Deck {
-  
-}
