+++ src/main/java/com/codedifferently/casino/Games/CardGame.java
@@ -1,0 +1,3 @@
+public class CardGame {
+  
+}
