+++ src/main/java/com/codedifferently/casino/Games/GoFish.java
@@ -1,0 +1,3 @@
+public class GoFish {
+  
+}
