--- src/main/java/com/codedifferently/casino/GoFish.java
@@ -1,3 +1,0 @@
-public class GoFish {
-  
-}
