+++ src/main/java/com/codedifferently/casino/Utilities/Card.java
@@ -1,0 +1,3 @@
+public class Card {
+  
+}
